Migrate the s3 plugin to be a metric sink.
