+++ sinks/s3/mock/s3.go
