+++ sinks/s3/testdata/aws/PutObject/2016/10/07/1475863542.json
