+++ sinks/s3/testdata/aws/PutObject/2016/10/13/1476370612.tsv.gz
