+++ sinks/s3/testdata/aws/PutObject/2016/10/14/1476481302.tsv.gz
