+++ util/csv.go
@@ -1,9 +1,11 @@
-package s3
+package util
 
 import (
+	"bytes"
+	"compress/gzip"
 	"encoding/csv"
-	"errors"
 	"fmt"
+	"io"
 	"strconv"
 	"strings"
 	"time"
@@ -48,12 +50,56 @@ var tsvSchema = [...]string{
 	TsvPartition:      "Partition",
 }
 
+// EncodeInterMetricsCSV returns a reader containing the gzipped CSV
+// representation of the InterMetric data, one row per InterMetric.
+// the AWS sdk requires seekable input, so we return a ReadSeeker here
+func EncodeInterMetricsCSV(
+	metrics []samplers.InterMetric, delimiter rune, includeHeaders bool,
+	hostname string, interval int,
+) (io.ReadSeeker, error) {
+	b := &bytes.Buffer{}
+	gzw := gzip.NewWriter(b)
+	w := csv.NewWriter(gzw)
+	w.Comma = delimiter
+
+	if includeHeaders {
+		// Write the headers first
+		headers := [...]string{
+			// the order here doesn't actually matter
+			// as long as the keys are right
+			TsvName:           TsvName.String(),
+			TsvTags:           TsvTags.String(),
+			TsvMetricType:     TsvMetricType.String(),
+			TsvInterval:       TsvInterval.String(),
+			TsvVeneurHostname: TsvVeneurHostname.String(),
+			TsvValue:          TsvValue.String(),
+			TsvTimestamp:      TsvTimestamp.String(),
+			TsvPartition:      TsvPartition.String(),
+		}
+
+		w.Write(headers[:])
+	}
+
+	// TODO avoid edge case at midnight
+	partitionDate := time.Now()
+	for _, metric := range metrics {
+		EncodeInterMetricCSV(metric, w, &partitionDate, hostname, interval)
+	}
+
+	w.Flush()
+	gzw.Close()
+	return bytes.NewReader(b.Bytes()), w.Error()
+}
+
 // EncodeInterMetricCSV generates a newline-terminated CSV row that describes
 // the data represented by the InterMetric.
 // The caller is responsible for setting w.Comma as the appropriate delimiter.
 // For performance, encodeCSV does not flush after every call; the caller is
 // expected to flush at the end of the operation cycle
-func EncodeInterMetricCSV(d samplers.InterMetric, w *csv.Writer, partitionDate *time.Time, hostName string, interval int) error {
+func EncodeInterMetricCSV(
+	d samplers.InterMetric, w *csv.Writer, partitionDate *time.Time,
+	hostName string, interval int,
+) error {
 	// TODO(aditya) some better error handling for this
 	// to guarantee that the result is proper JSON
 	tags := "{" + strings.Join(d.Tags, ",") + "}"
@@ -68,7 +114,7 @@ func EncodeInterMetricCSV(d samplers.InterMetric, w *csv.Writer, partitionDate *
 	case samplers.GaugeMetric:
 		metricType = "gauge"
 	default:
-		return errors.New(fmt.Sprintf("Encountered an unknown metric type %s", d.Type.String()))
+		return fmt.Errorf("encountered an unknown metric type %s", d.Type.String())
 	}
 
 	fields := [...]string{
@@ -94,7 +140,7 @@ func EncodeInterMetricCSV(d samplers.InterMetric, w *csv.Writer, partitionDate *
 // String returns the field Name.
 // eg tsvName.String() returns "Name"
 func (f tsvField) String() string {
-	return fmt.Sprintf(strings.Replace(tsvSchema[f], "tsv", "", 1))
+	return strings.Replace(tsvSchema[f], "tsv", "", 1)
 }
 
 // each key in tsvMapping is guaranteed to have a unique value
